Replace connect account service
